docker/django/requirements.txt
Django==3.2
django-jsonfield==1.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Pillow==7.1.1
PyJWT==1.7.1
psycopg2-binary==2.9.2
python-dotenv